Added readme, how_to_run.txt, fixed script typos.
