--- LOG_PATH_IS_UNDEFINED/LOG_FILE_IS_UNDEFINED
@@ -1,8 +1,0 @@
-2023-10-31 22:59:13.477 [main] INFO  o.s.t.c.s.AnnotationConfigContextLoaderUtils - Could not detect default configuration classes for test class [com.ramcel.cinema.reservation.service.unit.ScreeningServiceTest]: ScreeningServiceTest does not declare any static, non-private, non-final, nested classes annotated with @Configuration.
-2023-10-31 22:59:13.538 [main] INFO  o.s.b.t.c.SpringBootTestContextBootstrapper - Found @SpringBootConfiguration com.ramcel.cinema.reservation.Application for test class com.ramcel.cinema.reservation.service.unit.ScreeningServiceTest
-2023-10-31 23:00:31.832 [main] INFO  o.s.t.c.s.AnnotationConfigContextLoaderUtils - Could not detect default configuration classes for test class [com.ramcel.cinema.reservation.service.integration.TicketServiceIntegrationTest]: TicketServiceIntegrationTest does not declare any static, non-private, non-final, nested classes annotated with @Configuration.
-2023-10-31 23:00:31.909 [main] INFO  o.s.b.t.c.SpringBootTestContextBootstrapper - Found @SpringBootConfiguration com.ramcel.cinema.reservation.Application for test class com.ramcel.cinema.reservation.service.integration.TicketServiceIntegrationTest
-2023-10-31 23:00:50.475 [main] INFO  o.s.t.c.s.AnnotationConfigContextLoaderUtils - Could not detect default configuration classes for test class [com.ramcel.cinema.reservation.service.integration.TicketServiceIntegrationTest]: TicketServiceIntegrationTest does not declare any static, non-private, non-final, nested classes annotated with @Configuration.
-2023-10-31 23:00:50.537 [main] INFO  o.s.b.t.c.SpringBootTestContextBootstrapper - Found @SpringBootConfiguration com.ramcel.cinema.reservation.Application for test class com.ramcel.cinema.reservation.service.integration.TicketServiceIntegrationTest
-2023-10-31 23:06:03.714 [main] INFO  o.s.t.c.s.AnnotationConfigContextLoaderUtils - Could not detect default configuration classes for test class [com.ramcel.cinema.reservation.service.integration.ScreeningServiceIntegrationTest]: ScreeningServiceIntegrationTest does not declare any static, non-private, non-final, nested classes annotated with @Configuration.
-2023-10-31 23:06:03.787 [main] INFO  o.s.b.t.c.SpringBootTestContextBootstrapper - Found @SpringBootConfiguration com.ramcel.cinema.reservation.Application for test class com.ramcel.cinema.reservation.service.integration.ScreeningServiceIntegrationTest
